CT-1391 move FOI/SAR-specific bits out of Case::Base
Necessary as certain attributes and properties aren't shared between FOI and SAR
cases, including::

- validations
- enums
- after_create methods
- has_paper_trail declarations
