+++ app/models/case/base.rb
@@ -36,24 +36,6 @@ class Case::Base < ApplicationRecord
 
   default_scope { where( deleted: false) }
 
-  has_paper_trail only: %i{ name email received_date subject postal_address requester_type }
-
-
-  enum requester_type: {
-         academic_business_charity: 'academic_business_charity',
-         journalist: 'journalist',
-         member_of_the_public: 'member_of_the_public',
-         offender: 'offender',
-         solicitor: 'solicitor',
-         staff_judiciary: 'staff_judiciary',
-         what_do_they_know: 'what_do_they_know'
-       }
-
-  enum delivery_method: {
-         sent_by_post: 'sent_by_post',
-         sent_by_email: 'sent_by_email',
-       }
-
   jsonb_accessor :properties,
                  escalation_deadline: :date,
                  internal_deadline: :date,
@@ -139,18 +121,8 @@ class Case::Base < ApplicationRecord
 
   validates :name,presence: true
 
-  validates :email, presence: true, on: :create, if: -> { postal_address.blank? }
   validates :email, format: { with: /\A.+@.+\z/ }, if: -> { email.present? }
-  validates :postal_address,
-            presence: true,
-            on: :create,
-            if: -> { email.blank? || sent_by_post? }
-  validates :requester_type, :received_date, :delivery_method , presence: true
-  validates :message, presence: true, if: -> { sent_by_email? }
-  validates :uploaded_request_files,
-            presence: true,
-            on: :create,
-            if: -> { sent_by_post? }
+  validates_presence_of :received_date
   validates :subject, presence: true, length: { maximum: 100 }
   validates :type, presence: true, exclusion: { in: %w{Case}, message: "Case type can't be blank" }
 
@@ -268,7 +240,6 @@ class Case::Base < ApplicationRecord
                 :set_deadlines
   before_update :update_deadlines
   before_save :prevent_number_change
-  after_create :process_uploaded_request_files, if: :sent_by_post?
 
   delegate :available_events, to: :state_machine
