+++ app/views/cases/sar/_new_form_common.html.slim
@@ -5,7 +5,7 @@ section
   = f.text_field :subject_full_name
 
   = f.radio_button_fieldset :subject_type,
-                            choices: Case::SAR::VALID_SUBJECT_TYPES
+                            choices: Case::SAR::subject_types.keys
 
   #third-party
     = f.radio_button_fieldset :third_party
