+++ spec/models/case/base_spec.rb
@@ -36,9 +36,6 @@ RSpec.describe Case::Base, type: :model do
     build :case, received_date: Date.parse('16/11/2016')
   end
 
-  let(:no_postal)          { build :case, postal_address: nil             }
-  let(:no_postal_or_email) { build :case, postal_address: nil, email: nil }
-  let(:no_email)           { build :case, email: nil                      }
   let(:responding_team)    { create :responding_team                      }
   let(:responder)          { responding_team.responders.first             }
   let(:coworker)           { create :responder,
@@ -63,7 +60,6 @@ RSpec.describe Case::Base, type: :model do
 
   let(:kase) { create :case }
 
-
   describe 'has a factory' do
     it 'that produces a valid object by default' do
       Timecop.freeze non_trigger_foi.received_date + 1.day do
@@ -76,8 +72,6 @@ RSpec.describe Case::Base, type: :model do
     it { should validate_presence_of(:name)            }
     it { should validate_presence_of(:received_date)   }
     it { should validate_presence_of(:subject)         }
-    it { should validate_presence_of(:requester_type)  }
-    it { should validate_presence_of(:delivery_method) }
     it { should validate_presence_of(:type)            }
   end
 
@@ -371,49 +365,6 @@ RSpec.describe Case::Base, type: :model do
       expect(email_foi).to be_valid
       email_foi.message = nil
       expect(email_foi).not_to be_valid
-    end
-  end
-
-  describe 'enums' do
-    it do
-      should have_enum(:requester_type).
-        with_values(
-          [
-            'academic_business_charity',
-            'journalist',
-            'member_of_the_public',
-            'offender',
-            'solicitor',
-            'staff_judiciary',
-            'what_do_they_know'
-          ]
-        )
-
-      should have_enum(:delivery_method).
-          with_values(
-              [
-                  'sent_by_email',
-                  'sent_by_post'
-              ]
-          )
-    end
-  end
-
-  context 'without a postal or email address' do
-    it 'is invalid' do
-      expect(no_postal_or_email).not_to be_valid
-    end
-  end
-
-  context 'without a postal_address' do
-    it 'is valid with an email address' do
-      expect(no_postal).to be_valid
-    end
-  end
-
-  context 'without an email address' do
-    it 'is valid with a postal address' do
-      expect(no_email).to be_valid
     end
   end
 
@@ -1156,36 +1107,6 @@ RSpec.describe Case::Base, type: :model do
   describe 'search' do
     it 'returns case with a number that matches the query' do
       expect(Case::Base.search(accepted_case.number)).to match_array [accepted_case]
-    end
-  end
-
-  describe 'papertrail versioning', versioning: true do
-
-    before(:each) do
-      @kase = create :case, name: 'aaa', email: 'aa@moj.com', received_date: Date.today, subject: 'subject A', postal_address: '10 High Street', requester_type: 'journalist'
-      @kase.update!(name: 'bbb', email: 'bb@moj.com', received_date: 1.day.ago, subject: 'subject B', postal_address: '20 Low Street', requester_type: 'offender')
-    end
-
-    it 'saves all values in the versions object hash' do
-      version_hash = YAML.load(@kase.versions.last.object)
-      expect(version_hash['email']).to eq 'aa@moj.com'
-      expect(version_hash['received_date']).to eq Date.today
-      expect(version_hash['subject']).to eq 'subject A'
-      expect(version_hash['postal_address']).to eq '10 High Street'
-      expect(version_hash['requester_type']).to eq 'journalist'
-    end
-
-    it 'can reconsititue a record from a version (except for received_date)' do
-      original_kase = @kase.versions.last.reify
-      expect(original_kase.email).to eq 'aa@moj.com'
-      expect(original_kase.subject).to eq 'subject A'
-      expect(original_kase.postal_address).to eq '10 High Street'
-      expect(original_kase.requester_type).to eq 'journalist'
-    end
-
-    it 'does not reconstitute the received date properly because of an interaction with govuk_date_fields' do
-      original_kase = @kase.versions.last.reify
-      expect(original_kase.received_date).to eq 1.day.ago.to_date
     end
   end
